requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
python-dotenv==1.0.0
datadog-api-client==2.28.0
APScheduler==3.10.4
psycopg2-binary==2.9.9
gunicorn==21.2.0
requests==2.31.0
